+++ ecs/filter_gen.go
@@ -11,7 +11,7 @@ package ecs
 type Filter0 struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -84,7 +84,7 @@ func (f *Filter0) Exclusive() *Filter0 {
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter0) Relations(rel ...Relation) *Filter0 {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -121,7 +121,7 @@ func (f *Filter0) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter0.Batch] or [Filter0.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter0) Query(rel ...Relation) Query0 {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 0)
 
@@ -167,7 +167,7 @@ func (f *Filter0) Query(rel ...Relation) Query0 {
 // Otherwise, changes to the origin filter or calls to [Filter0.Batch] or [Filter0.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter0) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -197,7 +197,7 @@ func (f *Filter0) checkModify() {
 type Filter1[A any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -272,7 +272,7 @@ func (f *Filter1[A]) Exclusive() *Filter1[A] {
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter1[A]) Relations(rel ...Relation) *Filter1[A] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -309,7 +309,7 @@ func (f *Filter1[A]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter1.Batch] or [Filter1.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter1[A]) Query(rel ...Relation) Query1[A] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 1)
 		for i := range 1 {
@@ -356,7 +356,7 @@ func (f *Filter1[A]) Query(rel ...Relation) Query1[A] {
 // Otherwise, changes to the origin filter or calls to [Filter1.Batch] or [Filter1.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter1[A]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -384,7 +384,7 @@ func (f *Filter1[A]) checkModify() {
 type Filter2[A any, B any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -458,7 +458,7 @@ func (f *Filter2[A, B]) Exclusive() *Filter2[A, B] {
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter2[A, B]) Relations(rel ...Relation) *Filter2[A, B] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -495,7 +495,7 @@ func (f *Filter2[A, B]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter2.Batch] or [Filter2.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter2[A, B]) Query(rel ...Relation) Query2[A, B] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 2)
 		for i := range 2 {
@@ -542,7 +542,7 @@ func (f *Filter2[A, B]) Query(rel ...Relation) Query2[A, B] {
 // Otherwise, changes to the origin filter or calls to [Filter2.Batch] or [Filter2.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter2[A, B]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -572,7 +572,7 @@ func (f *Filter2[A, B]) checkModify() {
 type Filter3[A any, B any, C any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -649,7 +649,7 @@ func (f *Filter3[A, B, C]) Exclusive() *Filter3[A, B, C] {
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter3[A, B, C]) Relations(rel ...Relation) *Filter3[A, B, C] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -686,7 +686,7 @@ func (f *Filter3[A, B, C]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter3.Batch] or [Filter3.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter3[A, B, C]) Query(rel ...Relation) Query3[A, B, C] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 3)
 		for i := range 3 {
@@ -733,7 +733,7 @@ func (f *Filter3[A, B, C]) Query(rel ...Relation) Query3[A, B, C] {
 // Otherwise, changes to the origin filter or calls to [Filter3.Batch] or [Filter3.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter3[A, B, C]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -763,7 +763,7 @@ func (f *Filter3[A, B, C]) checkModify() {
 type Filter4[A any, B any, C any, D any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -841,7 +841,7 @@ func (f *Filter4[A, B, C, D]) Exclusive() *Filter4[A, B, C, D] {
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter4[A, B, C, D]) Relations(rel ...Relation) *Filter4[A, B, C, D] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -878,7 +878,7 @@ func (f *Filter4[A, B, C, D]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter4.Batch] or [Filter4.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter4[A, B, C, D]) Query(rel ...Relation) Query4[A, B, C, D] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 4)
 		for i := range 4 {
@@ -925,7 +925,7 @@ func (f *Filter4[A, B, C, D]) Query(rel ...Relation) Query4[A, B, C, D] {
 // Otherwise, changes to the origin filter or calls to [Filter4.Batch] or [Filter4.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter4[A, B, C, D]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -955,7 +955,7 @@ func (f *Filter4[A, B, C, D]) checkModify() {
 type Filter5[A any, B any, C any, D any, E any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -1034,7 +1034,7 @@ func (f *Filter5[A, B, C, D, E]) Exclusive() *Filter5[A, B, C, D, E] {
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter5[A, B, C, D, E]) Relations(rel ...Relation) *Filter5[A, B, C, D, E] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -1071,7 +1071,7 @@ func (f *Filter5[A, B, C, D, E]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter5.Batch] or [Filter5.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter5[A, B, C, D, E]) Query(rel ...Relation) Query5[A, B, C, D, E] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 5)
 		for i := range 5 {
@@ -1118,7 +1118,7 @@ func (f *Filter5[A, B, C, D, E]) Query(rel ...Relation) Query5[A, B, C, D, E] {
 // Otherwise, changes to the origin filter or calls to [Filter5.Batch] or [Filter5.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter5[A, B, C, D, E]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -1148,7 +1148,7 @@ func (f *Filter5[A, B, C, D, E]) checkModify() {
 type Filter6[A any, B any, C any, D any, E any, F any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -1228,7 +1228,7 @@ func (f *Filter6[A, B, C, D, E, F]) Exclusive() *Filter6[A, B, C, D, E, F] {
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter6[A, B, C, D, E, F]) Relations(rel ...Relation) *Filter6[A, B, C, D, E, F] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -1265,7 +1265,7 @@ func (f *Filter6[A, B, C, D, E, F]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter6.Batch] or [Filter6.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter6[A, B, C, D, E, F]) Query(rel ...Relation) Query6[A, B, C, D, E, F] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 6)
 		for i := range 6 {
@@ -1312,7 +1312,7 @@ func (f *Filter6[A, B, C, D, E, F]) Query(rel ...Relation) Query6[A, B, C, D, E,
 // Otherwise, changes to the origin filter or calls to [Filter6.Batch] or [Filter6.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter6[A, B, C, D, E, F]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -1342,7 +1342,7 @@ func (f *Filter6[A, B, C, D, E, F]) checkModify() {
 type Filter7[A any, B any, C any, D any, E any, F any, G any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -1423,7 +1423,7 @@ func (f *Filter7[A, B, C, D, E, F, G]) Exclusive() *Filter7[A, B, C, D, E, F, G]
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter7[A, B, C, D, E, F, G]) Relations(rel ...Relation) *Filter7[A, B, C, D, E, F, G] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -1460,7 +1460,7 @@ func (f *Filter7[A, B, C, D, E, F, G]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter7.Batch] or [Filter7.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter7[A, B, C, D, E, F, G]) Query(rel ...Relation) Query7[A, B, C, D, E, F, G] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 7)
 		for i := range 7 {
@@ -1507,7 +1507,7 @@ func (f *Filter7[A, B, C, D, E, F, G]) Query(rel ...Relation) Query7[A, B, C, D,
 // Otherwise, changes to the origin filter or calls to [Filter7.Batch] or [Filter7.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter7[A, B, C, D, E, F, G]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations
@@ -1537,7 +1537,7 @@ func (f *Filter7[A, B, C, D, E, F, G]) checkModify() {
 type Filter8[A any, B any, C any, D any, E any, F any, G any, H any] struct {
 	world        *World
 	ids          []ID
-	relations    []RelationID
+	relations    []relationID
 	components   []*componentStorage
 	filter       filter
 	cache        cacheID
@@ -1619,7 +1619,7 @@ func (f *Filter8[A, B, C, D, E, F, G, H]) Exclusive() *Filter8[A, B, C, D, E, F,
 // Can be called multiple times in chains, or once with multiple arguments.
 func (f *Filter8[A, B, C, D, E, F, G, H]) Relations(rel ...Relation) *Filter8[A, B, C, D, E, F, G, H] {
 	f.checkModify()
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	f.numRelations = uint8(len(f.relations))
 	return f
 }
@@ -1656,7 +1656,7 @@ func (f *Filter8[A, B, C, D, E, F, G, H]) Unregister() {
 // Otherwise, changes to the origin filter or calls to [Filter8.Batch] or [Filter8.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter8[A, B, C, D, E, F, G, H]) Query(rel ...Relation) Query8[A, B, C, D, E, F, G, H] {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	if f.components == nil {
 		f.components = make([]*componentStorage, 8)
 		for i := range 8 {
@@ -1703,7 +1703,7 @@ func (f *Filter8[A, B, C, D, E, F, G, H]) Query(rel ...Relation) Query8[A, B, C,
 // Otherwise, changes to the origin filter or calls to [Filter8.Batch] or [Filter8.Query]
 // with different relationship targets may modify stored instances.
 func (f *Filter8[A, B, C, D, E, F, G, H]) Batch(rel ...Relation) *Batch {
-	f.relations = relations(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations, f.numRelations)
+	f.relations = relationSlice(rel).toRelations(f.world, &f.filter.mask, f.ids, f.relations[:f.numRelations])
 	var start uint8
 	if f.cache != maxCacheID {
 		start = f.numRelations