+++ ecs/relation.go
@@ -15,80 +15,84 @@ var relationTp = reflect.TypeFor[RelationMarker]()
 // E.g. to iterate over all entities that are the child of a certain parent entity.
 type RelationMarker struct{}
 
-// Relation is the common interface for specifying relationship targets.
-// It is implemented by [Rel], [RelIdx] and [RelID].
+type relationID struct {
+	target    Entity
+	component ID
+}
+
+// Relation is the common type for specifying relationship targets.
+// It can be created with [Rel], [RelIdx] and [RelID].
 //
 //   - [Rel] is safe, but has some run-time overhead for component [ID] lookup.
 //   - [RelIdx] is fast but more error-prone.
 //   - [RelID] is used in the [Unsafe] API.
-type Relation interface {
-	id(ids []ID, world *World) ID
-	targetEntity() Entity
+type Relation struct {
+	target        Entity
+	component     ID
+	componentType reflect.Type
+	index         uint8
 }
 
-// RelationID specifies an entity relation target by component [ID].
-// Create with [RelID].
-//
-// It is used in Ark's unsafe, ID-based API.
-type RelationID struct {
-	component ID
-	target    Entity
+func (r *Relation) relationIDForUnsafe(world *World) relationID {
+	if r.index < 255 {
+		panic("relations created with RelIdx can't be used in the unsafe API, use RelID or Rel instead")
+	}
+	if r.componentType != nil {
+		r.component = TypeID(world, r.componentType)
+		r.componentType = nil
+	}
+	return relationID{
+		target:    r.target,
+		component: r.component,
+	}
+}
+
+func relID(id ID, target Entity) relationID {
+	return relationID{
+		target:    target,
+		component: id,
+	}
+}
+
+// id returns the component ID of this RelationID.
+func (r *Relation) id(ids []ID, world *World) ID {
+	if r.index < 255 {
+		return ids[r.index]
+	}
+	if r.componentType != nil {
+		r.component = TypeID(world, r.componentType)
+		r.componentType = nil
+	}
+	return r.component
+}
+
+// targetEntity returns the target [Entity] of this RelationID.
+func (r *Relation) targetEntity() Entity {
+	return r.target
 }
 
 // RelID creates a new [Relation] for a component ID.
 //
 // It is used in Ark's unsafe, ID-based API.
-func RelID(id ID, target Entity) RelationID {
-	return RelationID{
-		component: id,
+func RelID(id ID, target Entity) Relation {
+	return Relation{
 		target:    target,
+		component: id,
+		index:     255,
 	}
-}
-
-// id returns the component ID of this RelationID.
-func (r RelationID) id(ids []ID, world *World) ID {
-	return r.component
-}
-
-// targetEntity returns the target [Entity] of this RelationID.
-func (r RelationID) targetEntity() Entity {
-	return r.target
-}
-
-// relationType specifies an entity relation target by component type.
-// Create with [Rel].
-//
-// It can be used as a safer but slower alternative to [relationIndex].
-type relationType[C any] struct {
-	target Entity
 }
 
 // Rel creates a new [Relation] for a component type.
 //
 // It can be used as a safer but slower alternative to [RelIdx].
+// Required a component ID lookup when used the first time.
+// Un reuse, it is as fast as [RelIdx].
 func Rel[C any](target Entity) Relation {
-	return relationType[C]{
-		target: target,
+	return Relation{
+		target:        target,
+		componentType: reflect.TypeFor[C](),
+		index:         255,
 	}
-}
-
-// id returns the component ID of this RelationID.
-func (r relationType[C]) id(ids []ID, world *World) ID {
-	return ComponentID[C](world)
-}
-
-// targetEntity returns the target [Entity] of this RelationID.
-func (r relationType[C]) targetEntity() Entity {
-	return r.target
-}
-
-// relationIndex specifies an entity relation target by component index.
-// Create with [RelIdx].
-//
-// It can be used as faster but more error-prone alternative to [Rel].
-type relationIndex struct {
-	index  uint8
-	target Entity
 }
 
 // RelIdx creates a new [Relation] for a component index.
@@ -98,32 +102,33 @@ type relationIndex struct {
 // Note that the index refers to the position of the component in the generics
 // of e.g. a [Map2] or [Filter2].
 // This should not be confused with component [ID] as obtained by [ComponentID]!
+// For component IDs, use [RelationID].
 func RelIdx(index int, target Entity) Relation {
-	return relationIndex{
+	return Relation{
 		index:  uint8(index),
 		target: target,
 	}
 }
 
-// id returns the component ID of this RelationIndex.
-func (r relationIndex) id(ids []ID, world *World) ID {
-	return ids[r.index]
-}
+// Helper for converting relationSlice
+type relationSlice []Relation
 
-// targetEntity returns the target [Entity] of this RelationIndex.
-func (r relationIndex) targetEntity() Entity {
-	return r.target
-}
-
-// Helper for converting relations
-type relations []Relation
-
-func (r relations) toRelations(world *World, mask *bitMask, ids []ID, out []RelationID, startIdx uint8) []RelationID {
-	// TODO: can this be made more efficient?
-	out = out[:startIdx]
+func (r relationSlice) toRelations(world *World, mask *bitMask, ids []ID, out []relationID) []relationID {
 	if len(r) == 0 {
 		return out
 	}
+	// Fast special case for a single relation (20% speedup)
+	if len(r) == 1 {
+		rel := r[0]
+		id := rel.id(ids, world)
+		world.storage.checkRelationTarget(rel.targetEntity())
+		world.storage.checkRelationComponent(id)
+		if !mask.Get(id) {
+			panic(fmt.Sprintf("requested relation component with ID %d was not specified in the filter or map", id.id))
+		}
+		return append(out, relationID{target: rel.target, component: id})
+	}
+	// Slower with loop for more than one relation
 	for _, rel := range r {
 		id := rel.id(ids, world)
 		world.storage.checkRelationTarget(rel.targetEntity())
@@ -131,29 +136,30 @@ func (r relations) toRelations(world *World, mask *bitMask, ids []ID, out []Rela
 		if !mask.Get(id) {
 			panic(fmt.Sprintf("requested relation component with ID %d was not specified in the filter or map", id.id))
 		}
-		out = append(out, RelationID{
-			component: id,
-			target:    rel.targetEntity(),
-		})
+		out = append(out, relationID{target: rel.target, component: id})
 	}
 	return out
 }
 
-func (e Entity) toRelation(world *World, id ID, out []RelationID) []RelationID {
+func (r relationSlice) toRelationIDsForUnsafe(world *World, out []relationID) []relationID {
+	for _, rel := range r {
+		out = append(out, rel.relationIDForUnsafe(world))
+	}
+	return out
+}
+
+func (e Entity) toRelation(world *World, id ID, out []relationID) []relationID {
 	world.storage.checkRelationTarget(e)
 	world.storage.checkRelationComponent(id)
 	out = out[:0]
-	out = append(out, RelationID{
-		component: id,
-		target:    e,
-	})
+	out = append(out, relationID{target: e, component: id})
 	return out
 }
 
 // Helper for converting relations
 type relationEntities []Entity
 
-func (r relationEntities) toRelation(world *World, id ID, out []RelationID) []RelationID {
+func (r relationEntities) toRelation(world *World, id ID, out []relationID) []relationID {
 	out = out[:0]
 	if len(r) == 0 {
 		return out
@@ -161,10 +167,7 @@ func (r relationEntities) toRelation(world *World, id ID, out []RelationID) []Re
 	for _, rel := range r {
 		world.storage.checkRelationTarget(rel)
 		world.storage.checkRelationComponent(id)
-		out = append(out, RelationID{
-			component: id,
-			target:    rel,
-		})
+		out = append(out, relationID{target: rel, component: id})
 	}
 	return out
 }