+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [[unpublished]](https://github.com/mlange-42/ark/compare/v0.5.2...main)
+
+### Breaking changed
+
+- Converts interface `Relation` to a concrete struct type (#306)
+- Removes `RelationID`, use `Relation` instead (#306)
+
+### Performance
+
+- Speeds up relation handling by making `Relation` interface a concrete type (#306)
+
 ## [[v0.5.2]](https://github.com/mlange-42/ark/compare/v0.5.1...v0.5.2)
 
 ### Performance